Разделели один стейт на два независимых и раскидали по редьюсерам
